Move wait_for_sun_alt to keep pocs out of utils
